requirements.txt
python-telegram-bot==13.15
apscheduler
google-generativeai
newsapi-python
pytz